# src/main/java/com/example/flightsearchapi/ApiDocumantation.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/flights/update/{flightId}": {
      "put": {
        "tags": [
          "flight-controller"
        ],
        "operationId": "updateFlight",
        "parameters": [
          {
            "name": "flightId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Flight"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Flight"
                }
              }
            }
          }
        }
      }
    },
    "/airports/{id}": {
      "get": {
        "tags": [
          "airport-controller"
        ],
        "operationId": "getAirportById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "airport-controller"
        ],
        "operationId": "updateAirport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Airport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "airport-controller"
        ],
        "operationId": "deleteAirport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/token": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "token",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/flights/add": {
      "post": {
        "tags": [
          "flight-controller"
        ],
        "operationId": "addFlight",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Flight"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Flight"
                }
              }
            }
          }
        }
      }
    },
    "/airports": {
      "get": {
        "tags": [
          "airport-controller"
        ],
        "operationId": "getAllAirports",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Airport"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "airport-controller"
        ],
        "operationId": "addAirport",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Airport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          }
        }
      }
    },
    "/flights/{flightId}": {
      "get": {
        "tags": [
          "flight-controller"
        ],
        "operationId": "getFlightById",
        "parameters": [
          {
            "name": "flightId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Flight"
                }
              }
            }
          }
        }
      }
    },
    "/flights/trigger": {
      "get": {
        "tags": [
          "flight-controller"
        ],
        "operationId": "triggerFlight",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/flights/search": {
      "get": {
        "tags": [
          "flight-controller"
        ],
        "operationId": "searchFlights",
        "parameters": [
          {
            "name": "departureAirport",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Airport"
            }
          },
          {
            "name": "arrivalAirport",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Airport"
            }
          },
          {
            "name": "departureDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "returnDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Flight"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/flights/all": {
      "get": {
        "tags": [
          "flight-controller"
        ],
        "operationId": "getAllFlights",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Flight"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/flights/delete/{flightId}": {
      "delete": {
        "tags": [
          "flight-controller"
        ],
        "operationId": "deleteFlight",
        "parameters": [
          {
            "name": "flightId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Airport": {
        "required": [
          "city"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "city": {
            "maxLength": 45,
            "minLength": 1,
            "type": "string"
          }
        }
      },
      "Flight": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "departureAirport": {
            "$ref": "#/components/schemas/Airport"
          },
          "arrivalAirport": {
            "$ref": "#/components/schemas/Airport"
          },
          "departureDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "returnDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  }
}